==> _00_Annual-Update/old code/02_multivariate_ordinations.rmd <==
---
title: "Multivariate Analyses for Fish and Inverts"
author: Nick Toimieri
date: sys.date()
output: 
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    number_sections: true
    fig_caption: true
fontsize: 11pt

---

```{r SETUP, include=FALSE}

data_year = 2024

# A bunch of initial set up junk

rm(list = ls())

##### knitr stuff and libraries ####
source("R-functions-ocnms.r")

# Paths for files
home_dir = getwd()
# home_dir = "C:/Users/nick.tolimieri/Documents/GitHub/OCNMS/"

fig_dir = paste0(home_dir,"/figures/")
# data_dir = paste0(home_dir,"/data/")
data_dir = paste0("~/GitHub/OCNMS/Data/",data_year,"/")
results_dir = paste0(home_dir,"/results/")
other_dir = paste0(home_dir,"/other/")


# tools in use 
library(knitr)
# library(tidyr)
# library(dplyr)
library(tidyverse)
library(stringr)
library(tinytex) # for pdfs
library(RColorBrewer)
# display.brewer.all(colorblindFriendly = TRUE)
library(readxl)

# stats packages etc
library(vegan)
library(BiodiversityR)
# library(pracma)
# library(factoextra)

# install.packages("remotes")
# for Bray curtis with no spp in common.
# remotes::install_github("phytomosaic/ecole")
# library(ecole)

# options
knitr::opts_chunk$set(
  fig.path = fig_dir,
  cache.path = 'cache/graphics-', 
  echo=FALSE,
  error=FALSE,
  include = TRUE,
  dev='png',
  dpi=300,
  warning=FALSE,
  #out.width = '100%',
  fig.align='center'
  
  )

par(ps=10, cex=1)

# load some external functions
# source("R-functions-ocnms.r")

```

# Triggers and settings

```{r Triggers-and-Settings}

# bring in settings from previous univariate figure or set
# maintain consistency between files

settings = readRDS('settings.rds')

min.vis = as.numeric(settings['min.vis'])
years = as.numeric(unlist(settings['years']))
site.col = data.frame(settings['site.col'])
colnames(site.col) = c('site','col')
sites = site.col$site
year.pch = data.frame(settings['year.pch'])
colnames(year.pch) = c('year','pch','col')

# Set common stuff for ordinations here ####
# Options are: NA, "sqrt", "4th-root","log"
# for 'log' the transform is actually log(x+1), need the log+1 because of zeros.

kelp.transform = 'sqrt'
fish.transform = 'sqrt'
invert.transform = 'sqrt'
nperm  = 999

# Some csv files for naming later

fish_codes = data.frame(read.csv( "spp_codes_fish.csv" ))
swath_codes = data.frame(read.csv( "spp_codes_swath.csv"))
kelp_codes = data.frame(read.csv( "spp_codes_kelp.csv") )

```


# Import & Prepare Data for Ordinations

## Kelp

```{r import-data}

# import data separate fish and swath files ####
# transect level data ####

file.cols = c("year" ,"site", "transect","area", "zone")

fish0 <- readRDS( paste0(data_dir,  "Data_FISH_taxa_wide.rds") )
kelp0 <- readRDS( paste0(data_dir,  'Data_KELP_wide.rds') )
invert0 <- readRDS( paste0(data_dir,  "Data_INVERT_wide.rds") )

# select species, years, sites, & visibility ####
# delete 2015 data

fish = fish0[ fish0$year > 2015 &
               fish0$site %in% sites,]

kelp  = kelp0[kelp0$year > 2015 &
               kelp0$site %in% sites,
               ]

inverts  = invert0[invert0$year > 2015 &
                   invert0$site %in% sites,] 
inverts = inverts %>% rename(`hermit crabs` = hermit_crabs)

# add in some plotting information

fish$col = site.col$col[ match(fish$site,site.col$site)]
kelp$col = site.col$col[ match(kelp$site,site.col$site)]
inverts$col = site.col$col[ match(inverts$site,site.col$site)]

fish$pch = year.pch$pch[match(fish$year, year.pch$year) ]
kelp$pch = year.pch$pch[match(kelp$year, year.pch$year) ]
inverts$pch = year.pch$pch[match(inverts$year, year.pch$year) ]

```

```{r final-prep}
# fix some names for nicer plots
kelp.depth = fish.depth = invert.depth = '_5&10m_'
fish <- fish %>% rename('lingcod' = 'OPEL') %>%
                 rename('greenlings' = 'HEXA') %>%
                 rename('copper.rockfish' = 'SECA') %>%
                 rename('cabazon' = 'SCMA') %>%
                 rename('china.rockfish' = 'SENE') %>%
                 rename('black.rockfish' = 'SEME')
fish.spp = c('lingcod', 'greenlings','copper.rockfish','cabazon','china.rockfish','black.rockfish')
write.csv(fish.spp, paste0(fig_dir,"FishSpp-for-CAP.csv"))


kelp <- kelp %>% rename('Macrocystis' = 'MACPYR') %>%
                 rename('Nereocystis' = 'NERLUE') %>%
                 rename('Pterygophora' = 'PTECAL')%>% 
                  mutate(Other = sum(ALAMAR+COSCOS+CYMTRI+DESSPP+LAMSET+SACGRO+PLEGAR ))

kelp.spp = c('Macrocystis','Nereocystis','Pterygophora','Other')

invert.taxa = c("anemone", 
                "blood star",
                "brood star", 
                "chiton", 
                "crabs", 
                "cucumber",
                "green urchin", 
                "hermit crabs",
                "kelp crab",
                "large star",
                "leather star" , 
                "medium star", 
                "nudibranch",
                "Pisaster" , 
                "purple urchin" ,
                "Pycnopodia",
                "red urchin", 
                "shelled gastropod",
                "sponge", 
                "tunicate")
# invert.taxa = str_replace(invert.taxa, " ", '.')

write.csv(invert.taxa, paste0(fig_dir,"InvertSpp-for-CAP.csv"))

yoy.spp =    c("SECAyoy", "SEPIyoy", "SEMYyoy", "SEBYTyoy","RYOY")
write.csv(yoy.spp, paste0(fig_dir,"YoYSpp-for-CAP.csv"))

# need regular data frames or the column titles mess with the ordination software
kelp <-    data.frame(kelp)
fish <-    data.frame(fish)
#inverts <- data.frame(inverts %>% rename('hermit.crab' = 'hermit_crabs'))
#invert.taxa[invert.taxa == "hermit_crabs"] <- 'hermit.crab'

```

# Unconstrained Ordinations

non-metric MDS just to have a look. Kind of ugly. No alibi. Both depths.  

```{r nMDS-plots, fig.width=6, fig.height=6}

run.nmds = TRUE

if(run.nmds == TRUE){
mds_run_plot <- function( original.file, species,title, legend.pos){
  # remove rows with NA.
  df = na.omit(original.file)
  mds_data = df[,species]
  z = rowSums(mds_data)
  mds_data = mds_data[ z != 0,]
  df1 = df[z != 0,]
  mds.fit = metaMDS( mds_data,  distance = 'bray')

  xscores = scores(mds.fit)$sites
  df1 = cbind(df1,xscores)

  XLIM =c(min(df1$NMDS1),max(df1$NMDS1))
  YLIM =c(min(df1$NMDS2),max(df1$NMDS2))

  ordiplot(mds.fit, display = 'species', type = 'text', xlim=XLIM, ylim = YLIM)
  
  points(df1$NMDS1,df1$NMDS2, pch=as.numeric(df1$pch), col=df1$col, bg=df1$col, cex=0.5)
  legend(legend.pos, legend = title, bty='n')
  legend(legend.pos, legend = paste0('Stress = ',round(mds.fit$stress,2)), 
         inset=c(0,0.1), bty='n', cex=0.8)
  
  return(df1)
}


par( mfrow = c(2,2), ps=10, cex=1, pty='s', mar=c(4,4,0,0))

df0 = kelp[, c('site','year','zone','pch','col',kelp.spp)]
mds_kelp = mds_run_plot( df0, kelp.spp, 'Algae','topright' )

df0 = inverts[, c('site','year','zone','pch','col',invert.taxa)]
mds_invert = mds_run_plot( df0, invert.taxa, 'Inverts','topleft' )

df0 = fish[, c('site','year','zone','pch','col',fish.spp)]
mds_fish = mds_run_plot( df0, fish.spp, 'Fish','topleft' )

df0 = fish[, c('site','year','zone','pch','col',yoy.spp)]
mds_yoy = mds_run_plot( df0, yoy.spp, 'Sebastes YOY','bottomleft' )

}

```

# Constrained Ordinations

I ran things twice:

1. Once including all depths (5 & 10 m)
2. Once including only 5 m.

## Ordination - 5 & 10 m

```{r ordinations, include = FALSE}
## these analyses are run at the transect level
## later graphing may summarize them at different scales

kelp_depth = fish_depth = invert_depth = '_5&10m_'
nperm = 999
run.ord = TRUE

if(run.ord == TRUE){
  print("Working on kelp ordination")
    kelp_ord = run.multivar.nt( data.file = kelp, drop2015 = TRUE, nperm=nperm,
                        spp = kelp.spp, both_zones = TRUE,
                        data.transform = kelp.transform,
                        outname = paste0( results_dir, "Kelp-")
                        )
  saveRDS(kelp_ord,file = paste0(results_dir,"Kelp-ordination-n",nperm,"-",kelp.transform,kelp_depth,".rds"))

  print("Working on fish ordination")
  fish_ord = run.multivar.nt( data.file = fish, drop2015 = TRUE, nperm=nperm,
                        spp = fish.spp, both_zones = TRUE,
                        data.transform = fish.transform,
                        outname = paste0( results_dir, "Fish-")
                        )
  saveRDS(fish_ord,file = paste0(results_dir,"Fish-ordination-n",nperm,"-",fish.transform,fish_depth,".rds"))

  print("Working on Sebastes YOY ordination")
  yoy_ord = run.multivar.nt( data.file = fish,  drop2015 = TRUE, nperm=nperm,
                        spp = yoy.spp, both_zones = TRUE,
                        data.transform = fish.transform,
                        outname = paste0( results_dir, "YOY-")
                        )
  saveRDS(yoy_ord,file = paste0(results_dir,"YOY-ordination-n",nperm,"-",fish.transform,fish_depth,".rds"))

  print("Working on invert ordination")
  invert_ord = run.multivar.nt( data.file = inverts, drop2015 = TRUE, nperm=nperm,
                        spp = invert.taxa, both_zones = TRUE,
                        data.transform = invert.transform,
                        outname = paste0( results_dir, "Inverts-")
                        )
  saveRDS(invert_ord,file = paste0(results_dir,"Invert-ordination-n",nperm,"-",invert.transform,invert_depth,".rds"))

}else{ # reload previously run files to save time
     kelp_ord    <- readRDS(
       paste0(results_dir,"Kelp-ordination-n",nperm,"-",kelp.transform, kelp_depth,".rds"))
     fish_ord   <- readRDS( 
       paste0(results_dir,"Fish-ordination-n",nperm,"-",fish.transform, fish_depth,".rds"))
     yoy_ord    <- readRDS(
       paste0(results_dir,"YOY-ordination-n",nperm,"-",fish.transform, fish_depth,".rds"))
     invert_ord <- readRDS( 
       paste0(results_dir,"Invert-ordination-n",nperm,"-",invert.transform, invert_depth,".rds"))
}

```

### Combined ordination plot

NOTE to self: These plots use `r nperm` permutations. Double check that these are the ones you want.

Kelp transform = `r kelp.transform`
Invert transform = `r invert.transform`
Fish transform = `r fish.transform`
YOY transform = `r fish.transform`

Note if 'log', then log(x+1) because of all the zeros.

Separate Sites and Species plots for clarity.

Note: spp score must be min of abs(0.2) to show up on plot. So only black and blue show up on yoy plot.

```{r Figure-3-v2, fig.width=4.13, fig.height=8}

par( mfrow = c(4,2), pty='s' , mar = c(3.5,2,0,0), mgp=c(2,1,0))

xy=c(-7,7)
xy2 = c(-1.7, 1.7)
var.ylim = c(0,0.8)
var.cex = 0.9
plotspp = FALSE
sppseparate = TRUE
sppcol = 'black'

min.score = 0.2
scores.cex = 0.9

p_kelp = ifelse( kelp_ord$cap$manova$stats[1,6]<0.001 , "p < 0.001",
         paste0("p < ", round(kelp_ord$cap$manova$stats[1,6],3)))

p_inv = ifelse( invert_ord$cap$manova$stats[1,6]<0.001 , "p < 0.001",
         paste0("p < ", round(invert_ord$cap$manova$stats[1,6],3)))

p_fish = ifelse( fish_ord$cap$manova$stats[1,6]<0.001 , "p < 0.001",
         paste0("p < ", round(fish_ord$cap$manova$stats[1,6],3)))

p_yoy = ifelse( yoy_ord$cap$manova$stats[1,6]<0.001 , "p < 0.001",
         paste0("p < ", round(yoy_ord$cap$manova$stats[1,6],3)))



rownames(invert_ord$cap$cproj) <- str_replace( rownames(invert_ord$cap$cproj), "[.]" , " ")
rownames(fish_ord$cap$cproj) <- str_replace( rownames(fish_ord$cap$cproj), "[.]" , " ")

### fix names for ordination ####

# fix yoy rownames
rn <- rownames(yoy_ord$cap$cproj)
rn[rn == 'SECAyoy'] <- 'CQB'
rn[rn == 'SEPIyoy'] <- 'Canary'
rn[rn == 'SEMYyoy'] <- 'Blue RF'
rn[rn == 'Blue'] <- 'Blue RF'
rn[rn == 'SEBYTyoy'] <- 'YTB'
rn[rn == 'RYOY'] <- 'Juv RF'
rownames(yoy_ord$cap$cproj) <- rn

# fix kelp names
rn <- rownames(kelp_ord$cap$cproj)
rn[rn == 'Nerocystis'] <- 'Nereocystis'
rn[rn == 'Pterogophora'] <- 'Pterygophora'
rownames(kelp_ord$cap$cproj) <- rn

# fix fish names
rn <- rownames(fish_ord$cap$cproj)
rn[rn == 'blue rockfish'] <- 'Blue RF'
rn[rn == 'china rockfish'] <- 'China RF'
rn[rn == 'copper rockfish'] <- 'Copper RF'
rn[rn == 'cabazon'] <- 'Cabezon'
rn[rn == 'black rockfish'] <- 'Black RF'
rn[rn == 'lingcod'] <- 'Lingcod'
rn[rn == 'greenlings'] <- 'Greenling'

rownames(fish_ord$cap$cproj) <- rn

# fix invert names
rn <- rownames(invert_ord$cap$cproj)
rn <- str_to_title(rn)
rn[rn == 'Pynopodia'] <- 'Pycnopodia'
# rn[rn == 'Purple Urchin'] <- 'PU'
# rn[rn == 'Red Urchin'] <- 'RU'
# rn[rn == 'Green Urchin'] <- 'GU'
# rn[rn == 'Large Star'] <- 'LS'
# rn[rn == 'Chiton'] <- 'CH'


sppcol2 = rep('black',18)
sppcol2[c(5,11,12)] = 'red'

rownames(invert_ord$cap$cproj) <- rn

inset = c(-0.05 ,0)
yr = year.pch[year.pch$year != 2015 & year.pch$year!=2020 ,]
# kelps
Plot_Ordination( data.file = kelp_ord$cap.df , ord.file = kelp_ord$cap, min.score = min.score,
                 Yform = 'LD', Xform =  ' ~ site + year + col + pch', pval = p_kelp,
                 Xlim = c(-3,3), Ylim=c(-3,3), Xlim2 = xy2, Ylim2=xy2, scores.cex = scores.cex, scores.font = 3,
                 plot.species = plotspp , spp.separate = sppseparate, 
                 fig.lab = 'a) Kelp', fig.legend=TRUE, legend.inset = c(-0.02,0.05),
                 sppcol = sppcol)

#legend('bottomleft', legend=site.col$site, pch=19, col=site.col$col, bty='n', cex=0.8)


# legend('topright', legend=yr$year, pch=as.numeric(yr$pch), col='darkgrey',pt.bg = 'black', bty='n', cex=0.8 )
legend('topleft',legend = 'b)', bty='n', inset = inset)

# invets
Plot_Ordination( data.file = invert_ord$cap.df , ord.file = invert_ord$cap, min.score = min.score, pval = p_inv, 
                 Yform = 'LD', Xform =  ' ~ site + year + col + pch',scores.cex = scores.cex,
                 Xlim = xy, Ylim=c(-4,4),  Xlim2 = c(-1.3,1), Ylim2=c(-0.5,1),
                 plot.species = FALSE , spp.separate = sppseparate, fig.lab = 'c) Inverts', sppcol = sppcol2)

# z = data.frame(invert_ord$cap$cproj [ c('purple urchin','red urchin','green urchin'),])
# z$spp = rownames(z)
# text( z$LD1 , z$LD2, z$spp, col='red')
legend('topleft',legend = 'd)', bty='n', inset = inset)
# fish
Plot_Ordination( data.file = fish_ord$cap.df , ord.file = fish_ord$cap, min.score = min.score,
                 Yform = 'LD', Xform =  ' ~ site + year + col + pch', pval = p_fish,
                 Xlim = c(-2,2), Ylim=c(-2,2), Xlim2 = c(-1.7,0.25), Ylim2=c(-1,1),scores.cex = scores.cex,
                 plot.species = plotspp , spp.separate = sppseparate, fig.lab = 'e) Fishes', sppcol = sppcol)


legend('topleft',legend = 'f)', bty='n', inset = inset)
# sebastes yoy
Plot_Ordination( data.file = yoy_ord$cap.df , ord.file = yoy_ord$cap, min.score = min.score,
                 Yform = 'LD', Xform =  ' ~ site + year + col + pch', pval = p_yoy,
                 Xlim = c(-2,5), Ylim=c(-2.25, 2.25), Xlim2 = c(-0.5, 1), Ylim2=c(-0.4,0.2), scores.cex = scores.cex,
                 plot.species = plotspp , spp.separate = sppseparate, fig.lab = 'g) Juvenile rockfishes', sppcol = sppcol)

# legend('bottomleft', legend=yr$year, pch=as.numeric(yr$pch), col='darkgrey',pt.bg = 'black', bty='n', cex=0.8 )
legend('topleft',legend = 'h)', bty='n', inset = inset)






```

### Explained variance figure YSD

``` {r Figure-4, fig.width=3.18, fig.height=3}

axislabels = c('D', "S",'Y','SxD','DxY','SxY','Res')

f = data.frame(cbind('Fish' , axislabels, fish_ord$pm[1:7,'R2']))
s = data.frame(cbind('Rockfish YOY', axislabels,yoy_ord$pm[1:7,'R2']))
k = data.frame(cbind('Kelp', axislabels,kelp_ord$pm[1:7,'R2']))
v = data.frame(cbind('Inverts', axislabels,invert_ord$pm[1:7,'R2']))

df = rbind(f,s,k,v)
colnames(df) <- c('Spp','x', 'r2')
df$r2 = as.numeric(df$r2)

df$Spp = as.character(df$Spp)
df$Spp[df$Spp =='Rockfish YOY'] <- 'Juv. rockfishes'

df$Spp <- factor(df$Spp, levels = c("Kelp","Inverts","Fish",'Juv. rockfishes'))

df$x <-   factor(df$x , levels = axislabels)



hist.col = c(RColorBrewer::brewer.pal(12,'Paired')[c(3,2,6,7)])

library(ggplot2)
p = ggplot( df , aes(x = x,  y = r2, fill=Spp)) +
    # geom_point() + 
    #geom_jitter(size = 4) +
    geom_bar(stat="identity", position=position_dodge(), color='black') +
    scale_fill_manual( values =  hist.col )+
    ylab('Prop. explained variance') +
    xlab("Fixed terms") +
    theme_bw() + 
    theme( legend.title = element_blank(),
           legend.text = element_text(size = 8),
           legend.key.size = unit(0.7,'lines'),
           legend.position = c(0.25,0.83),
           legend.background = element_blank(),
           axis.text = element_text(size = 8))


p

```

## Ordinations 5-m depth only

```{r ordinations-5m, include = FALSE, eval=FALSE}
## these analyses are run at the transect level
## later graphing may summarize them at different scales

# get just 5 m data

kelp.depth = fish.depth = invert.depth = '_5-m_'

kelp5   = kelp %>% filter(zone == 5)
fish5   = fish %>% filter(zone == 5)
invert5 = inverts %>% filter(zone == 5)


run.ord = FALSE

if(run.ord == TRUE){
  print("Working on kelp ordination")
    kelp_ord = run.multivar.nt( data.file = kelp, drop2015 = TRUE, nperm=nperm,
                        spp = kelp.spp, both_zones = FALSE,
                        data.transform = kelp.transform,
                        outname = paste0( results_dir, "Kelp-")
                        )
  saveRDS(kelp_ord,file = paste0(results_dir,"Kelp-ordination-n",nperm,"-",kelp.transform,kelp.depth,".rds"))

  print("Working on fish ordination")
  fish_ord = run.multivar.nt( data.file = fish, drop2015 = TRUE, nperm=nperm,
                        spp = fish.spp, both_zones = FALSE,
                        data.transform = fish.transform,
                        outname = paste0( results_dir, "Fish-")
                        )
  saveRDS(fish_ord,file = paste0(results_dir,"Fish-ordination-n",nperm,"-",fish.transform,fish.depth,".rds"))

  print("Working on Sebastes YOY ordination")
  yoy_ord = run.multivar.nt( data.file = fish,  drop2015 = TRUE, nperm=nperm,
                        spp = yoy.spp, both_zones = FALSE,
                        data.transform = fish.transform,
                        outname = paste0( results_dir, "YOY-")
                        )
  saveRDS(yoy_ord,file = paste0(results_dir,"YOY-ordination-n",nperm,"-",fish.transform,fish.depth,".rds"))

  print("Working on invert ordination")
  invert_ord = run.multivar.nt( data.file = inverts, drop2015 = TRUE, nperm=nperm,
                        spp = invert.taxa, both_zones = FALSE,
                        data.transform = invert.transform,
                        outname = paste0( results_dir, "Inverts-")
                        )
  saveRDS(invert_ord,file = paste0(results_dir,"Invert-ordination-n",nperm,"-",invert.transform,invert.depth,".rds"))

}else{ # reload previously run files to save time
     kelp_ord    <- readRDS(
       paste0(results_dir,"Kelp-ordination-n",nperm,"-",kelp.transform,kelp.depth,".rds"))
     fish_ord   <- readRDS( 
       paste0(results_dir,"Fish-ordination-n",nperm,"-",fish.transform,fish.depth,".rds"))
     yoy_ord    <- readRDS(
       paste0(results_dir,"YOY-ordination-n",nperm,"-",fish.transform,fish.depth,".rds"))
     invert_ord <- readRDS( 
       paste0(results_dir,"Invert-ordination-n",nperm,"-",invert.transform,invert.depth,".rds"))
}

```

### Combined ordination plot 5 m

NOTE to self: These plots use `r nperm` permutations. Double check that these are the ones you want.

Kelp transform = `r kelp.transform`
Invert transform = `r invert.transform`
Fish transform = `r fish.transform`
YOY transform = `r fish.transform`

Note if 'log', then log(x+1) because of all the zeros.


Separate Sites and Species plots for clarity.



```{r ordinations-plots-separate-spp-panes-5m, fig.width=6, fig.height=8, eval=FALSE}

par( mfrow = c(4,2), pty='s' , mar = c(2,4,0,0) )

xy=c(-7,7)
xy2 = c(-1.2, 1.2)
var.ylim = c(0,0.8)
var.cex = 0.7
plotspp = FALSE
sppseparate = TRUE
sppcol = 'black'


p_kelp = ifelse( kelp_ord$cap$manova$stats[1,6]<0.001 , "p < 0.001",
         paste0("p < ", round(kelp_ord$cap$manova$stats[1,6],3)))

p_inv = ifelse( invert_ord$cap$manova$stats[1,6]<0.001 , "p < 0.001",
         paste0("p < ", round(invert_ord$cap$manova$stats[1,6],3)))

p_fish = ifelse( fish_ord$cap$manova$stats[1,6]<0.001 , "p < 0.001",
         paste0("p < ", round(fish_ord$cap$manova$stats[1,6],3)))

p_yoy = ifelse( yoy_ord$cap$manova$stats[1,6]<0.001 , "p < 0.001",
         paste0("p < ", round(yoy_ord$cap$manova$stats[1,6],3)))

# fix some names
rownames(invert_ord$cap$cproj) <- str_replace( rownames(invert_ord$cap$cproj), "[.]" , " ")
rownames(fish_ord$cap$cproj) <- str_replace( rownames(fish_ord$cap$cproj), "[.]" , " ")


Plot_Ordination( data.file = kelp_ord$cap.df , ord.file = kelp_ord$cap, min.score = 0.2,
                 Yform = 'LD', Xform =  ' ~ site + year + col + pch', pval = p_kelp,
                 Xlim = c(-3,3), Ylim=c(-3,3), Xlim2 = xy2, Ylim2=xy2,
                 plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Kelp', sppcol = sppcol)

legend('topright', legend=site.col$site, pch=19, col=site.col$col, bty='n', cex=0.8 )
yr = year.pch[year.pch$year != 2015 & year.pch$year!=2020 ,]

legend('topleft', legend=yr$year, pch=as.numeric(yr$pch), col='darkgrey',pt.bg = 'black', bty='n', cex=0.8 )

# invets
Plot_Ordination( data.file = invert_ord$cap.df , ord.file = invert_ord$cap, min.score = 0.2, pval = p_inv, 
                 Yform = 'LD', Xform =  ' ~ site + year + col + pch',
                 Xlim = xy, Ylim=xy, Xlim2 = xy2, Ylim2=xy2,
                 plot.species = FALSE , spp.separate = sppseparate, fig.legend = 'Inverts', sppcol = sppcol)

# spp_scores = data.frame(invert_ord$cap$cproj[c('green urchin',
#                                                'red urchin',
#                                                'purple urchin', 
#                                                'leather star',
#                                                'brood star',
#                                                'Pisaster'),])
# text(spp_scores$LD1, spp_scores$LD2,rownames(spp_scores), col = 'red', cex=0.8)

# fish
Plot_Ordination( data.file = fish_ord$cap.df , ord.file = fish_ord$cap, min.score = 0.2,
                 Yform = 'LD', Xform =  ' ~ site + year + col + pch', pval = p_fish,
                 Xlim = c(-3,3), Ylim=c(-3,3), Xlim2 = xy2, Ylim2=xy2,
                 plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Fish', sppcol = sppcol)



# sebastes
Plot_Ordination( data.file = yoy_ord$cap.df , ord.file = yoy_ord$cap, min.score = 0.2,
                 Yform = 'LD', Xform =  ' ~ site + year + col + pch', pval = p_yoy,
                 Xlim = c(-5,5), Ylim=c(-5, 5), Xlim2 = xy2, Ylim2=xy2,
                 plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Sebastes YOY', sppcol = sppcol)



```

### Explained variance figure 5 m

``` {r explained-variance-5m, fig.width=4.5, fig.height=3.5 , eval = FALSE}

axislabels = c("S",'Y','SY','Res')

f = data.frame(cbind('Fish' , axislabels, fish_ord$pm[1:4,'R2']))
s = data.frame(cbind('Sebastes YOY', axislabels,yoy_ord$pm[1:4,'R2']))
k = data.frame(cbind('Kelp', axislabels,kelp_ord$pm[1:4,'R2']))
v = data.frame(cbind('Inverts', axislabels,invert_ord$pm[1:4,'R2']))

df = rbind(f,s,k,v)
colnames(df) <- c('Spp','x', 'r2')
df$r2 = as.numeric(df$r2)
df$Spp <- factor(df$Spp, levels = c("Kelp","Inverts","Fish",'Sebastes YOY'))

df$x <-   factor(df$x , levels = axislabels)

hist.col = c(RColorBrewer::brewer.pal(12,'Paired')[c(3,2,6,7)])

library(ggplot2)
p = ggplot( df , aes(x = x,  y = r2, fill=Spp)) +
    # geom_point() +
    #geom_jitter(size = 4) +
    geom_bar(stat="identity", position=position_dodge(), color='black') +
    scale_fill_manual( values =  hist.col )+
    ylab('Prop. explained variance') +
    xlab("") +
     theme_minimal()

p

```


# Ordinations Fish vs Kelp, YOY vs kelp, Inverts vs Kelp

## Just Adult Fishes

```{r set-depth}
# set.depth.5m = FALSE
# 
# if(set.depth.5m == TRUE){print('This analysis includes only 5-m depth.')}else{
#       print('This analysis includes 5 & 10 m') }
#       
```

```{r fish-kelp-capscale}
# distance-based redundancy analysis

# df0 = data.frame(readRDS(paste0(data_dir,'Data_Fish_Kelp_area_wide.rds')))
# df0 = df0 %>% filter(year != 2015)
# 
# # fish.spp = set above
# # kelp.spp = set above
# 
# 
#  ##### presently turned off ########## fish.spp doesn't match.
# df1 = df0[,c('site','year','zone','area',fish.spp,kelp.spp)]
# df1$id = paste(df1$site, df1$year, df1$zone, df1$area, sep='_')
# rownames(df1) <- df1$id
# 
# # drop rows with NA's -- for now. Check that kelp aren't just no kelp at depth.
# df = na.omit(df1)
# if(set.depth.5m == TRUE){df = df[df$zone == 5,]}
# 
# transform = fish.transform # to allow different transform if desired.
# dfx = df[,c(fish.spp)]
# cap.fish = matrix.transform(dfx,  transform )
# # zero-adjusted Bray_Curtis Similarity; Clarke et al 2006
# # Clarke, K.R., P.J. Somerfield, and M.G. Chapman. 2006. On resemblance measures for ecological studies, including taxonomic dissim and a zero-adjusted Bray–Curtis coefficient for denuded assemblages. J Exp Marine Biol and Ecol 330:55–80.
# # 
# # Smith, R.J. 2017. Solutions for loss of information in high-beta-diversity community data. Methods in Ecology and Evolution 8(1): 68-74.
# 
# fish.matrix =  ecole::bray0(cap.fish)
# 
# cap1 = capscale( fish.matrix ~ MACPYR + NERLUE + PTECAL , data = df , comm = dfx)
# fish_cap = anova(cap1)
# fish_pval = fish_cap[4]
# 
# par( ps=10, cex = 1, pty='s')
# ordiplot(cap1, xlim=c(-6,6), ylim=c(-6,6) )
# x = scores(cap1)
# df = cbind(df,x$sites)
# df$col = site.col$col[match(df$site,site.col$s)]
# points(df$CAP1,df$CAP2,pch=19, col=df$col)
# a = data.frame(x$species)
# text(a$CAP1, a$CAP2, rownames(a), col='red')
# 
# legend('topleft' , legend = site.col$site, pch=19, col = site.col$col, bty='n', cex=0.8)
# 
# fish.cap <- cap1
# fish.df <- df
# 
# saveRDS(list(cap1,df, fish_cap), file = paste0(results_dir,'Fish-Kelp-Ordination',transform,".rds") )
# 


```


## Just Sebastes YOY 


```{r yoy-kelp-capscale}
# distance-based redundancy analysis

# df0 = data.frame(readRDS(paste0(data_dir,'Data_Fish_Kelp_area_wide.rds')))
# df0 = df0 %>% filter(year != 2015)
# # fish.spp = set above
# # kelp.spp = set above
# 
# df1 = df0[,c('site','year','zone','area', yoy.spp,kelp.spp)]
# df1$id = paste(df1$site, df1$year, df1$zone, df1$area, sep='_')
# rownames(df1) <- df1$id
# 
# df = na.omit(df1)
# if(set.depth.5m == TRUE){df = df[df$zone == 5,]}
# dfx = df[,c(yoy.spp)]
# 
# transform = fish.transform # to allow different transform if desired.
# dfx = df[,c(yoy.spp)]
# yoy.fish = matrix.transform(dfx,  transform )
# 
# fish.matrix =  ecole::bray0(yoy.fish)
# 
# cap1 = capscale( fish.matrix ~ MACPYR + NERLUE + PTECAL , data = df, comm = dfx)
# yoy_cap <- anova(cap1)
# yoy_pval = yoy_cap[4]
# 
# par( ps=10, cex = 1, pty='s')
# ordiplot(cap1, xlim=c(-6,6), ylim=c(-6,6) )
# x = scores(cap1)
# df = cbind(df,x$sites)
# df$col = site.col$col[match(df$site,site.col$s)]
# points(df$CAP1,df$CAP2,pch=19, col=df$col)
# a = data.frame(x$species)
# text(a$CAP1, a$CAP2, rownames(a), col='red')
# 
# legend('topleft' , legend = site.col$site, pch=19, col = site.col$col, bty='n', cex=0.8)
# 
# yoy.cap <- cap1
# yoy.df <- df
# 
# saveRDS(list(cap1,df, yoy_cap), file = paste0(results_dir,'YOY-Kelp-Ordination',transform,".rds"))

```

## Swath Inverts

```{r invert-kelp}

# df = data.frame(readRDS( paste0( data_dir,"Data_Inverts_Kelp_transect_wide.rds" ) ))
# df = df %>% filter(year != 2015)
# df = na.omit(df)
# transform = invert.transform # to allow different transform if desired.
# 
# dfx = df[,invert.taxa]
# 
# cap.inv = matrix.transform(dfx,  transform )
# invert.matrix = ecole::bray0(cap.inv)
# 
# cap1 = capscale( invert.matrix ~ MACPYR + NERLUE + PTECAL , data = df, comm = dfx)
# inv_cap <- anova(cap1)
# inv_pval <- inv_cap[4]
# 
# lims = 4
# 
# par( ps=10, cex = 1, pty='s')
# ordiplot(cap1, xlim = c(-lims,lims) , ylim=c(-lims,lims) , cex = 0.2)
# x = scores(cap1  )
# df = cbind(df,x$sites)
# df$col = site.col$col[match(df$site,site.col$s)]
# points(df$CAP1,df$CAP2,pch=19, col=df$col, cex = 1)
# 
# a = data.frame(x$species)
# text(a$CAP1, a$CAP2, rownames(a), col='red')
# 
# legend('topleft' , legend = site.col$site, pch=19, col = site.col$col, bty='n', cex=0.8)
# 
# inv.cap <- cap1
# inv.df <- df
# 
# saveRDS(list(cap1,df, inv_cap), file = paste0(results_dir,'Invert-Kelp-Ordination.rds'))



```

### XXX vs Kelp plotted together.

```{r capscale-combined, fig.width=6, fig.height=8}

# par( mfrow = c(2,2), ps = 10, cex = 1, pty='s' , mar = c(2,4,0,0) )
# plotspp = TRUE
# sppseparate = FALSE
# sppcol = 'black'
# min.score = 0.2
# 
# fish.df$pch = yoy.df$pch = inv.df$pch = 19
# lims = 4
# xy2 = c(-2,2)
# 
# # year symbols
# 
# # year.pch$pch = c(9, 15,16,17,18,21,19)
# 
# 
# fish.df$pch = year.pch$pch[match(fish.df$year,year.pch$year)]
# yoy.df$pch =  year.pch$pch[match(yoy.df$year,year.pch$year)]
# inv.df$pch =  year.pch$pch[match(inv.df$year,year.pch$year)]
# 
# Plot_Ordination( data.file = fish.df , ord.file = fish.cap, min.score = min.score, method = 'capscale',
#                  Yform = 'CAP', Xform =  ' ~ site + year + col + pch', bg.equals.col = TRUE,
#                  Xlim = c(-lims,lims), Ylim=c(-lims, lims), Xlim2 = xy2, Ylim2=xy2,
#                  plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Fish', sppcol = sppcol)
# text(-3.5 , 2.8 , paste0('p = ', fish_pval[1,1]) , pos=4 )
# 
# Plot_Ordination( data.file = yoy.df , ord.file = yoy.cap, min.score = min.score, method = 'capscale',
#                  Yform = 'CAP', Xform =  ' ~ site + year +  col + pch', bg.equals.col = TRUE,
#                  Xlim = c(-5,5), Ylim=c(-5, 5), Xlim2 = xy2, Ylim2=xy2,
#                  plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Sebastes YOY', sppcol = sppcol)
# text(-4.3 , 3.5 , paste0('p = ', yoy_pval[1,1]) , pos=4 )
# 
# Plot_Ordination( data.file = inv.df , ord.file = inv.cap, min.score = min.score, method = 'capscale',
#                  Yform = 'CAP', Xform =  ' ~ site + year + col + pch', bg.equals.col = TRUE,
#                  Xlim = c(-lims,lims), Ylim=c(-lims, lims), Xlim2 = xy2, Ylim2=xy2,
#                  plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Invertebrates', sppcol = sppcol)
# text(-3.5 , 2.8 , paste0('p = ', inv_pval[1,1]) , pos=4 )
# 
# # legend in extra space
# plot(1:10 , 1:10, pch = "" , xlab = NA, ylab = NA, xaxt = 'n', yaxt = 'n' , bty='n')
# 
# year.legend = year.pch[year.pch$year != 2015 & year.pch$year!=2020,]
# legend('left' , legend= site.col$site, pch=19, col = site.col$col , bty = 'n')
# legend('right' , legend=year.legend$year, pch=as.numeric(year.legend$pch), col = 'black' , bty = 'n')



```

### Same data plotted differently.

```{r capscale-combined-v2, fig.width=6, fig.height=8}

# par( mfrow = c(3,2), ps = 10, cex = 1, pty='s' , mar = c(2,4,0,0) )
# plotspp = FALSE
# sppseparate = TRUE
# sppcol =   site.col$col[3] 
# comm.col = site.col$col[1]
# min.score = 0.2
# 
# fish.df$pch = yoy.df$pch = inv.df$pch = 19
# lims = 4
# xy2 = c(-2,2)
# 
# fish.df$pch = year.pch$pch[match(fish.df$year,year.pch$year)]
# yoy.df$pch =  year.pch$pch[match(yoy.df$year,year.pch$year)]
# inv.df$pch =  year.pch$pch[match(inv.df$year,year.pch$year)]
# 
# 
# Plot_Ordination( data.file = fish.df , ord.file = fish.cap, min.score = min.score, method = 'capscale',
#                  Yform = 'CAP', Xform =  ' ~ site + year + col + pch', bg.equals.col = TRUE, 
#                  pval = paste0('p = ', fish_pval[1,1]) , plot.comm.scores = TRUE, comm.col =comm.col,
#                  Xlim = c(-lims,lims), Ylim=c(-lims, lims), Xlim2 = xy2, Ylim2=xy2,
#                  plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Fish', sppcol = sppcol)
# 
# 
# Plot_Ordination( data.file = yoy.df , ord.file = yoy.cap, min.score = min.score, method = 'capscale',
#                  Yform = 'CAP', Xform =  ' ~ site + year +  col + pch', bg.equals.col = TRUE,
#                   pval = paste0('p = ', yoy_pval[1,1]) , plot.comm.scores = TRUE, comm.col =comm.col,
#                  Xlim = c(-5,5), Ylim=c(-5, 5), Xlim2 = xy2, Ylim2=xy2,
#                  plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Sebastes YOY', sppcol = sppcol)
# 
# Plot_Ordination( data.file = inv.df , ord.file = inv.cap, min.score = min.score, method = 'capscale',
#                  Yform = 'CAP', Xform =  ' ~ site + year + col + pch', bg.equals.col = TRUE,
#                   pval = paste0('p = ', inv_pval[1,1]) , plot.comm.scores = TRUE, comm.col =comm.col,
#                  Xlim = c(-lims,lims), Ylim=c(-lims, lims), Xlim2 = c(-4,4), Ylim2=c(-4,4),
#                  plot.species = plotspp , spp.separate = sppseparate, fig.legend = 'Invertebrates', sppcol = sppcol)

# 
# year.legend = year.pch[year.pch$year != 2015 & year.pch$year!=2020,]
# legend('left' , legend= site.col$site, pch=19, col = site.col$col , bty = 'n')
# legend('right' , legend=year.legend$year, pch=as.numeric(year.legend$pch), col = 'black' , bty = 'n')


```
